gpu_service/requirements.txt
absl-py==1.0.0
amqp==5.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
billiard==3.6.4.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
Flask==2.0.3
Flask-Cors==3.0.10
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
kombu==5.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.0
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.5.1
notebook==6.4.10
numpy==1.19.1
oauthlib==3.2.0
opencv-python==4.3.0.36
openslide-python==1.1.2
packaging==21.3
pandas==1.1.0
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.7.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.8
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.1.1
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.2
Send2Trash==1.8.0
Shapely==1.8.1.post1
six==1.16.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.6.0+cu92
torchvision==0.7.0+cu92
tornado==6.1
tqdm==4.48.0
traitlets==4.3.3
typing-extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
zipp==3.6.0
